requirements.txt
pydantic>=2.7.1,<3.0.0
python-multipart==0.0.9
python-socketio>=5.12.1,<6.0.0
grpcio>=1.71.0,<2.0.0
passlib>=1.7.4,<2.0.0
pyjwt>=2.10.1,<3.0.0
requests==2.32.2
aiohttp==3.9.5
httpx>=0.28.1,<1.0.0
openai>=1.66.3,<2.0.0
langgraph>=0.3.18,<1.0.0
langchain-core>=0.3.47,<1.0.0
langchain-openai>=0.3.9,<1.0.0
langchain>=0.3.23,<1.0.0
pandas>=2.2.3,<3.0.0
psycopg2>=2.9.10,<3.0.0
sqlalchemy>=2.0.40,<3.0.0
scipy>=1.15.2,<2.0.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
matplotlib>=3.8.0
seaborn>=0.13.0
gradio>=4.19.0
psycopg2
langgraph-checkpoint-postgres
langchain-postgres
psycopg-binary
asyncpg
aiofiles
itsdangerous 
